feat: rename atoms
