+++ src/klip/index.ts
@@ -1,1 +1,4 @@
 export * from './useKlipLogin';
+export * from './execute';
+export * from './utils';
+export * from './types';
