+++ src/recoil/klip/selectors.ts
@@ -1,10 +1,10 @@
 import { selector } from 'recoil';
-import { klipAuthAtom } from './atoms';
+import { klipRequestKeyAtom } from './atoms';
 
 export const getKlipQrcodeSelector = selector({
   key: 'auth/getKlipQrcode',
   get: ({ get }) =>
     `https://klipwallet.com/?target=/a2a?request_key=${
-      get(klipAuthAtom).requestKey
+      get(klipRequestKeyAtom).requestKey
     }`,
 });
